requirements.txt
fastapi==0.109.1
uvicorn==0.23.2
sqlmodel==0.0.8
pydantic==2.6.0
scikit-learn==1.3.3
numpy==1.26.2
python-dotenv==1.0.1
google-generativeai==0.2.0
